Move ImageBorderSegment to border.rs
